* Changed to long double, use l goniometric functions

+++ src/precession.c
@@ -40,8 +40,8 @@ void ln_get_equ_prec (struct ln_equ_posn * mean_position, double JD, struct ln_e
 	long double t, t2, t3, A, B, C, zeta, eta, theta, ra, dec, mean_ra, mean_dec;
 	
 	/* change original ra and dec to radians */
-	mean_ra = ln_deg_to_rad(mean_position->ra);
-	mean_dec = ln_deg_to_rad(mean_position->dec);
+	mean_ra = ln_deg_to_rad (mean_position->ra);
+	mean_dec = ln_deg_to_rad (mean_position->dec);
 
 	/* calc t, zeta, eta and theta for J2000.0 Equ 20.3 */
 	t = (JD - JD2000) / 36525.0;
@@ -51,26 +51,26 @@ void ln_get_equ_prec (struct ln_equ_posn * mean_position, double JD, struct ln_e
 	zeta = 2306.2181 * t + 0.30188 * t2 + 0.017998 * t3;
 	eta = 2306.2181 * t + 1.09468 * t2 + 0.041833 * t3;
 	theta = 2004.3109 * t - 0.42665 * t2 - 0.041833 * t3;
-	zeta = ln_deg_to_rad(zeta);
-	eta = ln_deg_to_rad(eta);
-	theta = ln_deg_to_rad(theta); 
+	zeta = ln_deg_to_rad (zeta);
+	eta = ln_deg_to_rad (eta);
+	theta = ln_deg_to_rad (theta); 
 
 	/* calc A,B,C equ 20.4 */
-	A = cos(mean_dec) * sin(mean_ra + zeta);
-	B = cos(theta) * cos(mean_dec) * cos(mean_ra + zeta) - sin(theta) * sin (mean_dec);
-	C = sin(theta) * cos (mean_dec) * cos(mean_ra + zeta) + cos(theta) * sin(mean_dec);
+	A = cosl (mean_dec) * sinl (mean_ra + zeta);
+	B = cosl (theta) * cosl (mean_dec) * cosl (mean_ra + zeta) - sinl (theta) * sinl (mean_dec);
+	C = sinl (theta) * cosl (mean_dec) * cosl (mean_ra + zeta) + cosl (theta) * sinl (mean_dec);
 	
-	ra = atan2 (A,B) + eta;
+	ra = atan2l (A,B) + eta;
 	
 	/* check for object near celestial pole */
 	if (mean_dec > (0.4 * M_PI) || mean_dec < (-0.4 * M_PI)) {
 		/* close to pole */
-		dec = acos(sqrt(A * A + B * B));
+		dec = acosl (sqrt(A * A + B * B));
 		if (mean_dec < 0.)
 		  dec *= -1; /* 0 <= acos() <= PI */
 	} else {
 		/* not close to pole */
-		dec = asin (C);
+		dec = asinl (C);
 	}
 
 	/* change to degrees */
@@ -95,13 +95,13 @@ void ln_get_equ_prec2 (struct ln_equ_posn * mean_position, double fromJD, double
 	long double t, t2, t3, A, B, C, zeta, eta, theta, ra, dec, mean_ra, mean_dec, T, T2;
 	
 	/* change original ra and dec to radians */
-	mean_ra = ln_deg_to_rad(mean_position->ra);
-	mean_dec = ln_deg_to_rad(mean_position->dec);
+	mean_ra = ln_deg_to_rad (mean_position->ra);
+	mean_dec = ln_deg_to_rad (mean_position->dec);
 
 	/* calc t, T, zeta, eta and theta Equ 20.2 */
-	T = (fromJD - JD2000) / 36525.0;
+	T = ((long double) (fromJD - JD2000)) / 36525.0;
 	T *= 1.0 / 3600.0;
-	t = (toJD - fromJD) / 36525.0;
+	t = ((long double) (toJD - fromJD)) / 36525.0;
 	t *= 1.0 / 3600.0;
 	T2 = T * T;
 	t2 = t * t;
@@ -109,26 +109,26 @@ void ln_get_equ_prec2 (struct ln_equ_posn * mean_position, double fromJD, double
 	zeta = (2306.2181 + 1.39656 * T - 0.000139 * T2) * t + (0.30188 - 0.000344 * T) * t2 + 0.017998 * t3;
 	eta = (2306.2181 + 1.39656 * T - 0.000139 * T2) * t + (1.09468 + 0.000066 * T) * t2 + 0.018203 * t3;
 	theta = (2004.3109 - 0.85330 * T - 0.000217 * T2) * t - (0.42665 + 0.000217 * T) * t2 - 0.041833 * t3;
-	zeta = ln_deg_to_rad(zeta);
-	eta = ln_deg_to_rad(eta);
-	theta = ln_deg_to_rad(theta); 
+	zeta = ln_deg_to_rad (zeta);
+	eta = ln_deg_to_rad (eta);
+	theta = ln_deg_to_rad (theta); 
 
 	/* calc A,B,C equ 20.4 */
-	A = cos(mean_dec) * sin(mean_ra + zeta);
-	B = cos(theta) * cos(mean_dec) * cos(mean_ra + zeta) - sin(theta) * sin (mean_dec);
-	C = sin(theta) * cos (mean_dec) * cos(mean_ra + zeta) + cos(theta) * sin(mean_dec);
+	A = cosl (mean_dec) * sinl (mean_ra + zeta);
+	B = cosl (theta) * cosl (mean_dec) * cosl (mean_ra + zeta) - sinl (theta) * sinl (mean_dec);
+	C = sinl (theta) * cosl (mean_dec) * cosl (mean_ra + zeta) + cosl (theta) * sinl (mean_dec);
 	
-	ra = atan2 (A,B) + eta;
+	ra = atan2l (A,B) + eta;
 	
 	/* check for object near celestial pole */
 	if (mean_dec > (0.4 * M_PI) || mean_dec < (-0.4 * M_PI)) {
 		/* close to pole */
-		dec = acos(sqrt(A * A + B * B));
+		dec = acosl (sqrt(A * A + B * B));
 		if (mean_dec < 0.)
 		  dec *= -1; /* 0 <= acos() <= PI */
 	} else {
 		/* not close to pole */
-		dec = asin (C);
+		dec = asinl (C);
 	}
 
 	/* change to degrees */
